Refactoring of converter tests
