+++ activiti-cloud-services-runtime-bundle/activiti-cloud-services-events/src/test/java/org/activiti/cloud/services/events/converter/ToCloudProcessRuntimeTimerEventsConverterTest.java
@@ -1,0 +1,90 @@
+/*
+ * Copyright 2018 Alfresco, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.activiti.cloud.services.events.converter;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.verify;
+import static org.mockito.MockitoAnnotations.initMocks;
+
+import org.activiti.api.process.model.payloads.TimerPayload;
+import org.activiti.api.runtime.event.impl.BPMNTimerFiredEventImpl;
+import org.activiti.api.runtime.event.impl.BPMNTimerScheduledEventImpl;
+import org.activiti.api.runtime.model.impl.BPMNTimerImpl;
+import org.activiti.cloud.api.model.shared.impl.events.CloudRuntimeEventImpl;
+import org.activiti.cloud.api.process.model.events.CloudBPMNTimerFiredEvent;
+import org.activiti.cloud.api.process.model.events.CloudBPMNTimerScheduledEvent;
+import org.junit.Before;
+import org.junit.Test;
+import org.mockito.ArgumentMatchers;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+
+public class ToCloudProcessRuntimeTimerEventsConverterTest {
+
+    @InjectMocks
+    private ToCloudProcessRuntimeEventConverter converter;
+
+    @Mock
+    private RuntimeBundleInfoAppender runtimeBundleInfoAppender;
+
+    @Before
+    public void setUp() {
+        initMocks(this);
+    }
+
+    @Test
+    public void shouldConvertBPMNTimerFiredEventToCloudBPMNTimerFiredEvent() {
+        //given
+        BPMNTimerImpl timer = new BPMNTimerImpl("entityId");
+        timer.setProcessInstanceId("procInstId");
+        timer.setProcessDefinitionId("procDefId");
+        TimerPayload timerPayload = new TimerPayload();
+
+        timer.setTimerPayload(timerPayload);
+        BPMNTimerFiredEventImpl timerFiredEvent = new BPMNTimerFiredEventImpl(timer);
+
+        //when
+        CloudBPMNTimerFiredEvent cloudEvent = converter.from(timerFiredEvent);
+        assertThat(cloudEvent.getEntity()).isEqualTo(timer);
+        assertThat(cloudEvent.getProcessDefinitionId()).isEqualTo("procDefId");
+        assertThat(cloudEvent.getProcessInstanceId()).isEqualTo("procInstId");
+
+        //then
+        verify(runtimeBundleInfoAppender).appendRuntimeBundleInfoTo(ArgumentMatchers.any(CloudRuntimeEventImpl.class));
+    }
+    
+    @Test
+    public void shouldConvertBPMNTimerScheduledEventToCloudBPMNTimerScheduledEvent() {
+        //given
+        BPMNTimerImpl timer = new BPMNTimerImpl("entityId");
+        timer.setProcessInstanceId("procInstId");
+        timer.setProcessDefinitionId("procDefId");
+        TimerPayload timerPayload = new TimerPayload();
+
+        timer.setTimerPayload(timerPayload);
+        BPMNTimerScheduledEventImpl timerFiredEvent = new BPMNTimerScheduledEventImpl(timer);
+
+        //when
+        CloudBPMNTimerScheduledEvent cloudEvent = converter.from(timerFiredEvent);
+        assertThat(cloudEvent.getEntity()).isEqualTo(timer);
+        assertThat(cloudEvent.getProcessDefinitionId()).isEqualTo("procDefId");
+        assertThat(cloudEvent.getProcessInstanceId()).isEqualTo("procInstId");
+
+        //then
+        verify(runtimeBundleInfoAppender).appendRuntimeBundleInfoTo(ArgumentMatchers.any(CloudRuntimeEventImpl.class));
+    }
+}
